Caller takes ownership in FPDFTextObj_GetFont
Previously font data would be inaccessible after closing the page that
contained the text object that had the font. This change makes the font
outlive the page but also requires the caller to call the FPDFFont_Close API
in order to close the font after usage.

Since FPDFTextObj_GetFont is marked as an experimental API, changing the API
in this way should be okay. Even if a user of PDFium updates and does not
add calls to FPDFFont_Close, the worst case scenario is some leaked memory,
which shouldn't be that bad. I'm unsure if the font will be automatically
closed upon closing the document as a whole.

Change-Id: I555143946aaef336921a2ed08069909fda5c5475