PA: Temporary option for new PartitionAlloc location.
To make the partition_alloc directory a standalone library, I am going to:

1. Move from files from:
   /base/allocator/partition_allocator/*
   toward
   /base/allocator/partition_allocator/src/partition_alloc/*

2. Make PartitionAlloc to declare its include_dirs.

3. PartitionAlloc dependants will be able put PartitionAlloc wherever
   they want (e.g. probably in their third_party/partition_alloc) and
   to use:
   #include "partition_alloc/..."
   instead of:
   #include "base/allocator/partition_allocator/src/partition_alloc/..."

Unfortunately, the partition_alloc library is already used by PDFium.
Step (1) would break it. To achieve step 1 without breaking pdfium, this
patch adds an option for chromium to set the location of PartitionAlloc.

This is temporary and will be removed after step 1 is completed in
Chrome.
It a kind of 3-step change.

Bug:chromuium:1467773
Change-Id: I9f312f6a47976659a1e792c663006f3a75ec7fca