[Skia] Fix issues for printing 24bpp images
When using `CFX_SkiaDeviceDriver` for rendering to a 24 bbp bitmap, it
creates an internal bitmap buffer for fast processing and saves the actual output bitmap as `m_pOriginalBitmap`. Keeping both the original
and the internal bitmaps up-to-date is the key to get the correct
printing result.

This CL improves the rendering process in the following ways:

1. For `CFX_SkiaDeviceDriver` only: Add
   `CFX_RenderDevice::SyncInternalBitmaps()` method to make the
   underlying `m_pDeviceDriver` transfer the rendering result from its
   internal buffer to the `CFX_RenderDevice` bitmap. This allows
   syncing without destroying the `m_pDeviceDriver`.

2. In `CPDF_DeviceBuffer::OutputToDevice()`, `pBuffer` is used for
   creating a `CFX_RenderDevice` already before it got passed into
   `GetBackground()` to be painted white. This means the "old"
   `CFX_RenderDevice`'s device driver contains an internal buffer
   that's out-of-date (without the white paint). To keep the existing
   `CFX_RenderDevice` buffers fresh, use it as a parameter of
   `GetBackground()` so that its buffer gets updated and we can avoid
   creating a new layer of `CFX_RenderDevice`.

Also, removes the pixel tests that are fixed by this CL from the
suppression list.

Bug: chromium:1470768, pdfium:2054
Change-Id: I8442664ae58dc528442b40a8503100f9501f5ea1