Avoid returning the wrong buffer in CPDF_DIB::GetBuffer()
When `CPDF_DIB::m_bColorKey` is set, GetBuffer() should not return
`m_pCachedBitmap->GetBuffer()`, as that buffer is used for internal
calculations and should not be exposed directly to callers. By returning
an empty span in GetBuffer() instead, that signals to callers that they
need to Realize() the bitmap, which then uses GetScanline() to correctly
render the bitmap.

Bug: chromium:1478366
Change-Id: I5509b9908f6f53501a5029a90bc0d5c61441814a