Rename CPDF_ImageLoader::HandleFailure() to Finish()
HandleFailure() is the wrong name for the method, as the
CPDF_ImageLoader's actions always end with "failure". Instead, rename
the "ret" variables in the class to "should_continue" to indicate what
is actually happening.

Note that "should_continue" can be false either due to image loading
finishing, or failing. The CPDF_ImageRenderer class that ultimately
calls into CPDF_ImageLoader is the one that determines success vs.
failure.

Change-Id: I338d295493a1ba2260583fbc6931554e1f9d9144