Simplify CFX_DIBitmap::ConvertColorScale()
Since ConvertColorScale()'s inputs are always either {0, 0xffffff}, or
vice versa, its implementation can be simplified. Since the RGB channels
are always the same, many channel variables go away. Since the
difference between the foreground and background is always 255, the
"difference * gray / 255" calculations can become either "gray" or
"negative gray".

Along the way, pull some variables out of for-loops, when the variable
value is independent of the loop. Also rename ConvertColorScale() to
ConvertToGrayscale() and remove the return value that callers do not
check.

Change-Id: Ifc1613f98e86cfbcb29f07996b5431a0498e275b