Remove unreachable code from CFX_DIBBase::ConvertTo()
CFX_DIBBase::ConvertTo() first handles the case where the source and
destination formats are the same. Then it proceeds to handle the `kArgb`
format, which is also the only format with an alpha channel. When the
destination format is `kArgb`, the only way for `pSrcAlpha` to be
non-null is for the source format to also be `kArgb`. Since this case
has also been handled, it cannot occur again. After removing this
unreachable code, the code to set `pSrcAlpha` also becomes useless.
Change-Id: I544c2dcacb156140f1cace64fe38c23fd3f69c2c