Fix dangling ptrs in embeddertest found by BRP dangling ptr detector.
These were missed by the PDFium probing-style detector because the
raw_ptr<> were never destroyed. LSAN could not find it either, since
the memory was not leaked, just the dangling references were not
cleaned up beforehand.

The reason is that we have C++ classes inheriting from C-style structs,
hence there can not be virtual destructors. When holding an unique_ptr<>
to the C-style struct, the subclass is never destructed.

Change-Id: I631b2e07670e1e09111ffa9f35e600cc42b15323