Fix incorrect font dictionary usage in GeneratePopupAP()
When GeneratePopupAP() calls CPDF_DocPageData::GetFont(), GetFont()
expects a font dictionary. GeneratePopupAP() is incorrectly passing in
the resource dictionary for all fonts. This ultimately causes the font
mapping code to ask the PDFium embedder for a font with no font name.
Fix this by generating the font dictionary and the font resources
dictionary separately. Thus making it easy to call GetFont() with the
correct dictionary.

Change-Id: I313411864f993e096cbd41f440791bdf821bf268