Extend PDFEditImgTest.NewImageObjLoadJpeg to write out a PDF
Change this test case to do more than just generate an in-memory bitmap.
After checking the in-memory bitmap is correct, insert the bitmap into
the page and make sure that renders correctly. Then write out the PDF
and verify the generated PDF renders correctly. Do the same in the
NewImageObjLoadJpegInline test case.

Change-Id: Ib48c050f2444f97fe84010dfe065d7441aa770f7