Change GetFileContents() test utility to return a vector
Instead of returning a std::unique_ptr, with a separate out-parameter
for the size, change GetFileContents() to just return a std::vector.
Because it is only used for reading data files in tests,
GetFileContents() should never have to read an empty file. Therefore, it
can use an empty vector to indicate failure, and not have to wrap the
vector in an optional.

Change-Id: I485551e2ed4c47e98f4535cf91bbf6bff1ca99bb