Allow ProcessCrossRefV5Entry() to overwrite existing entries
Currently in CPDF_Parser::ProcessCrossRefV5Entry(), if there already
exists an entry for a given object number, ProcessCrossRefV5Entry() will
never overwrite the existing object. Remove this check so newer cross
reference V5 entries take precedence when cross references parsing
starts with the oldest entry. Keep the check to prevent overwriting when
parsing starts with the newest cross reference.

With the issue fixed, remove the suppression for the bug_1484283.pdf
test case. The only other touch up is to add an expectation image for
bug_1484283.pdf rendering in Windows GDI mode. The output is not a pixel
perfect match with the existing expectation files.

Bug: chromium:1484283
Change-Id: Ib06474c2110f50468ae57fbdc945b39431f0b907