Avoid setting the private tif_fd field in struct tiff
CTiffContext::InitDecoder() calls a helper function which wraps
TIFFClientOpen() and sets the tif_fd field in the returned struct. This
is a private field that the libtiff caller should not be reaching into.
The field also appears to be of no use ot CTiffContext.

Clean up C-style casts along the way.

Bug: chromium:1501495
Change-Id: Ic42e7c7e1af68483814f72333fd8a32547fb900f