Extract surrogate fusing from PDF_DecodeText() into own separate function
For UTF-16 inputs, PDF_DecodeText() used to fuse surrogates in the
main processing loop.

The main processing loop does a bunch of unrelated things, so this
moves surrogate fusing into a separate function that's called after
the main loop.

Since surrogate fusion either keeps the string the same (if there are
no surrogates) or makes it shorter, it can be done in-place.

Doing this in multiple passes is easier to think about, at least for me.

Modifies the code that was added in
https://pdfium-review.googlesource.com/c/pdfium/+/107831

(I'd like to also factor out the escape code handling, and after that
it's hopefully easy to add support for UTF-8 text strings, which is
my actual goal.)

No intended behavior change.

Change-Id: I36d9405ce0865b097263852f5914cfe57e8c4878