Simplify PDF_DecodeText() and new helper functions
I kept the code as-is when moving it. Now that it's moved and untangled,
we can simplify the code a bit.

Every modified function is supposed to have exactly the same behavior
as it had before this change. They no longer write to a `uint16_t`
temporary, so they no longer truncate characters to 16-bits, but
the caller only passed in 16-bit data in both cases, so no effective
behavior change.

(A future change will call StripLanguageCodes() with the output of
an UTF-8 decoder, which can pass 32-bit characters. For that future
use-case, the change is a progression.)

Change-Id: If56afbadb8282b042f4411a5d1930543c67a197b