Fix non-standalone ASAN component build on windows.
The source_set("raw_ptr") is part of the component("base"), so for
non-standalone component builds, PDFium should make use of that
component. This avoids conflicts between PDFium's direct use of
a source_set and //base's use of the same, as //base seems to be
getting pulled into builds.

Otherwise, just depend on source_set("raw_ptr").

Bug: chromium:1505391
Change-Id: Iad0033265e856bf5ab1daefbde516ff74cf32c08