Add std::u16string FX_UTF16Encode(WideStringView)
...and use it to implement WideString::ToUTF16LE().

No behavior change for now, but if we decide that it's ok that
FPDF_WIDESTRING becomes UTF-16 with the platform-native endianness
instead of staying UTF-16LE everywhere, we can switch almost all
current callers of WideString::ToUTF16LE() to FX_UTF16Encode() and
then make WideString::ToUTF16LE() no longer add zero-termination.

Even if we don't end up doing that, this change here separates
the "split surrogates" and "encode as bytes" concerns and makes things
more readable, so it's valuable in itself.

Change-Id: I9f60a702083f2d8f0b2c5030e1d975d983e674c8