Add flag to disable blink gc plugin check
The plugin now has a check to prohibit stl collections of gced types and
members. Pdfiun intentionally uses such collections and thus can't use
the new check.
This CL adds a flag to BUILD.gn to explicitly disable this check such
that it will not be used in pdfium even when it is enabled by default.

Bug: chromium:1470687
Change-Id: I009068eca560bee4d7d937532f52bb94f0e9b9b1