Mark MultiplyAlpha() consistently take a float parameter
Some MultiplyAlpha() methods take a float, while others take an integer.
It is too easy to mix them up, so switch them all to take floats. Then
update the related call stacks to pass the float all the way through.
This way, the float value that initially comes from
CPDF_GeneralState::GetFillAlpha() traverses as-is all the way to the
RenderDeviceDriverIface implementation.

Change-Id: I60e5d775aaf3e2d04ab81400df3d172319f9e543