Add CFX_DIBBase::TakePalette()
Give CFX_DIBBase::SetPalette() callers a way to set the palette with a
move instead of a copy. Use it where appropriate, which requires
changing ScanlineComposerIface::SetInfo() to take a vector instead of a
span.

Additionally, change some cfx_dibbase.cpp helper functions to initialize
palettes directly, instead of allocating a vector of a fixed size, and
then copying into it.

Change-Id: I7af470a503c068c1ae7fd5ace6eb99156624e36d