Unshare CFX_DIBitmap::SetChannelFromBitmap()
Stop using SetChannelFromBitmap() as a common method shared by
SetAlphaFromBitmap() and SetRedFromBitmap(). This is in preparation for
simplifying the 2 caller methods based on their unique input
constraints. Along the way, rename the 2 callers to better describe
their roles, encapsulate them if appropriate, and remove the input
parameter if it is not useful.

Change-Id: I0b78522bc75edf4d275298fa19076aef4aabd924