Add pixel test for an edited PDF that regressed
Similar to testing/resources/pixel/bug_1484283.pdf, add another
handcrafted PDF that is a variant of testing/resources/rectangles.pdf.
This one starts out with size 200x350, but the edit changed the size to
200x300.

It should render with size 200x300 to match to expectation PNGs.
However, the open bug means they do not render correctly, thus the test
results are currently suppressed.

Bug: pdfium:2122
Change-Id: Iaf2da742e95589b0755fcf336bd8351b491e7246