Revert ProcessCrossRefV5Entry() overwrite change
This effectively reverts the code changes in
https://pdfium-review.googlesource.com/112932 and puts the
bug_1484283.pdf pixel test back in a suppressed state, while retaining
some of the code cleanups that do not affect the bug. In the reverted
state, bug_2122.pdf renders correctly again, so unsuppress that test
case.

A follow-up CL [1] will fix the bug_1484283.pdf pixel test in a
different manner. Do that separately to make the changes clearer.

[1] https://pdfium-review.googlesource.com/116093

Bug: chromium:1484283,pdfium:2122
Change-Id: I4dabeac00ec236458d20d29b5e65992c76b6eddc