Fix PDF loading when an initial xref stream references a xref table
In CPDF_Parser, LoadAllCrossRefV4() will try to call LoadCrossRefV5() if
LoadCrossRefV4() fails for secondary xref data. However,
LoadAllCrossRefV5() only calls LoadCrossRefV5(). As a result, if
LoadAllCrossRefV5() only knows how to handle cross reference streams,
and fails if it ever encounters a cross reference table. Update
LoadAllCrossRefV5() to call LoadCrossRefV4() if LoadCrossRefV5() fails.
Then unsupress the pixel test that renders correctly as a result of this
fix.

Along the way, fix some nits for how CPDF_Dictionary variables in
CPDF_Parser are named and assigned to.

Bug: pdfium:2123
Change-Id: Id77056b459dcd58864e34821940e73a29caed927