Consolidate byteorder code
Remove third_party/base/sys_byteorder.h and pull out the bits that are
being used to core/fxcrt/byteorder.h. This removes an unnecessary layer
of complexity, as most of third_party/base/sys_byteorder.h is not even
used, and the remaining code will eventually be replaced with C++23
std::byteswap().

Bug: pdfium:2127
Change-Id: Ibf9b2e2ad674e9c0f651f844cc588a37e4c8c19a