Move {Byte,Wide}String::Trim{,Left,Right}() into StringTemplate<>
Move the forms which take arguments, but leave the no-arg form in
the byte/wide string files to allow for potential differences in the
default characters to be stripped.

Rename the no-arg form to StripUsual*() to avoid having to use a
`using` statement to work around the C++ idiosyncrasies of subclass
method overload name hiding.

Change-Id: Ib5adba3b79e394e0ad7eaf0a528526bfc55aed3e