Add EmbedderTestEnvironment::ReleaseFontInfo()
Add a dedicated step for EmbedderTestEnvironment to release the use of
its font info, and thus giving the tests that need to call
FPDF_FreeDefaultSystemFontInfo() a chance to do so before TearDown().

In a follow-up CL, PartitionAllocator class instances will stop living
forever. Thus after TearDown(), it would be too late to call
FPDF_FreeDefaultSystemFontInfo(), as the font info was allocated inside
a destroyed partition.

Bug: pdfium:1876
Change-Id: Ie35d6ef33e84aa7656ad33aa6b59f5c2bf4e914e