Fix incorrect matrix outputs in CPDF_PageContentGenerator
When CPDF_PageContentGenerator generates content stream data for a page
object, it currently does not take the object's container's current
transformation matrix into account. As a result, the generated output
effectively has the CTM applied twice for page objects whose CTM is not
the identity matrix, as shown in recently added PDFEditImgTest cases.

Change CPDF_PageContentGenerator to take the CTM into account for images
and forms. Then update the related PDFEditImgTest cases so they pass
with the expected results. Add TODOs for CPDF_PageContentGenerator
methods that process text and paths, as there are no associated tests
for those.

Bug: pdfium:2132
Change-Id: I8d4c783ddee5c44dfeca32ec5575d228572e2417