Convert EXPECT_STREQ(..., ...c_str()) to EXPECT_EQ(..., ...).
The objective is to stop calling .c_str() except when absolutely
necessary, as extracting a c_str() from one of our string types
leave us vulnerable to, say, injected \0.

The gtest macros can handle these _EQ comparisons so long as
operator==() is defined for the string type. The risk is that
a bug in operator==() might propagate to other tests, but we
test this itself.

Change-Id: I9bdbd857c45b2356719a19560a30c35a80b2bff5