Roll third_party/skia/ 9832c7722..49f18efdf (331 commits)
https://skia.googlesource.com/skia.git/+log/9832c7722061..49f18efdfa73

Include extra SkEncodedInfo.cpp and SkPngCodec.cpp in Windows Skia port
sources to overcome a build failure for the component build.  This is
necessary due to the change in http://review.skia.org/808496, as these
files are not automatically included by the skia_codec_core build
source set.

Created with:
  roll-dep third_party/skia

Change-Id: Ic323f75c302ff77b94aaa9f0f6046fef9ff6b462