Fix wrong matrix outputs in CPDF_PageContentGenerator for path objects
When CPDF_PageContentGenerator generates content stream data for a path
page object, it currently does not take the object's container's current
transformation matrix into account. Fix this and update the test that
now renders correctly.

With the above fix, FPDFEditEmbedderTest.SaveAndRender fails. This is
because the SaveAndRender test case has a non-identity current
transformation matrix in the original PDF, and then
FPDFPage_InsertObject() added a page object without any knowledge of the
CTM. Given the behavior of FPDFPage_InsertObject() cannot change w.r.t
the CTM, the only way to fix this is to set an extra bit to
differentiate CPDF_PageObject instances that were parsed from the PDF,
vs. CPDF_PageObject instances that were synthesized using PDFium APIs.

Bug: pdfium:2132
Change-Id: Id9238dc075dae7843eb1ed5131d120784f66c192