Properly account for CTM changes
In the existing code, CPDF_PageObjectHolder only stored the final value
of the current transformation matrix. This is insufficient information
for PDF editing and saving, where a page can have multiple content
streams and only some of them have been modified and get regenerated.

To properly account for CTM changes from stream to stream, update
CPDF_StreamContentParser to store the CTM at the end of streams, when
handling "cm" and "Q" operators. CPDF_PageObjectHolder can then take
this data as `m_AllCTMs`, and stop storing `m_LastCTM`. With this
additional information, CPDF_PageContentGenerator can compensate for the
CTM when handling the modified streams. It can do so at the stream
level, and individual objects no longer need to take the CTM into
account, because each stream will start with the CTM set to the identity
matrix. At the end of each stream, the CTM will be restored to the value
expected by the next stream.

With the above changes, `CPDF_PageObject::m_bUsesCTM` is no longer
necessary, as objects no longer use the CTM. Remove it and associated
code. The FPDFEditPathEmbedderTest.AddPathToRectanglesWithLeakyCTM test
case also works as expected, so update the test expectation and remove
the TODO comment.

Bug: pdfium:2142
Change-Id: I95fab8dd5a06da19743f24f9e89007b4bb3c34c9