Adjust BUILD.gn files for path-based unsafe buffer warnings.
Prepare for using the clang plugin-based filtering mechanism for unsafe
unsafe buffer warnings. This is enabled automatically as appropriate in
build/config/clang/BUILD.gn (sync'ed from outside of pdfium repository
along with other tools). It is enabled on a directory-by-directory basis
based on some paths.txt file (location to be specified via a GN variable
following a build/ roll).

Change-Id: I030dc5d15b2b9ea0541af146bcce5d61df13ffee