Fix parsing of PDFs with multiple cross reference streams
The fix for crbug.com/40932618 removed the ability to follow a chain of
trailers in hybrid-reference files. While that met the spec, it
regressed other PDFs because there was no call to code that follows the
chain of cross reference streams in hybrid-reference files. Add that
missing code to fix the regression.

Bug: 332462378
Change-Id: I520994ec2273f608f63849a3abdd41005abc19df