Change FPDF_StructElement_Attr_* APIs to never return a reference
This applies the change to FPDF_StructElement_Attr_GetNumberValue() in
https://pdfium-review.googlesource.com/110850 to other APIs like
FPDF_StructElement_Attr_GetType() and
FPDF_StructElement_Attr_GetStringValue() for consistency. With this
change, FPDF_StructElement_Attr_GetType() will no longer return
FPDF_OBJECT_REFERENCE, and callers will know the actual type and be able
to determine the right API to call to fetch the data.

Change-Id: I12563ac522fc25b7d7b5b05b578d1af7ca161f63