Convert to std::array<> in CXFA_ContentLayoutProcessor.
This resolves unsafe buffer usage, and avoids the confusing syntax of
(&arg)[3] required  for passing a reference to an array rather than
an array of references. With the std::array<> syntax, confuision is
avoided because it is simpler to understand.

In the process, fix an instance where we were declaring an argument
as an array[3] but were passing as an actual a reference into the
middle of the array[3]. This worked in the past due to the way
array[3] degrades to array* when being passed.

Change-Id: Iabce74ebc1589df30b6efa05391d98309acf8160