Convert to std::array<> in CXFA_ContentLayoutProcessor.
This helps to resolve unsafe buffer usage in the future, and avoids
the confusing syntax of T (&arg)[3] required  for passing a reference
to an array rather than an array of references (which is prohibited).
With the std::array<> syntax, confusion is avoided.

In the process, fix an instance where we were declaring an argument
as an array[3] but then invoking a -> operator against it. This worked
in the past due to the way array[3] degrades to array*.

Pass a reference in one place rather than a pointer to keep things
even simpler.

Change-Id: Iabce74ebc1589df30b6efa05391d98309acf8160