Convert to std::array<> in CXFA_ContentLayoutProcessor.
This resolves unsafe buffer usage, and avoids the confusing syntax of
(&arg)[3] required  for passing a reference to an array rather than
an array of references. With the std::array<> syntax, confusion is
avoided.

In the process, fix an instance where we were declaring an argument
as an array[3] but were invoking a -> operator against it. This worked
in the past due to the way array[3] degrades to array*.

Pass a reference in one place rather than a pointer to keep things
even simpler.

Change-Id: Iabce74ebc1589df30b6efa05391d98309acf8160