Tolerate extra JPEG2000 image channels in CPDF_DIB::LoadJpxBitmap()
JPEG2000 images can have more color channels than the number of color
components. Thus checking for exactly 4 channels may be too strict. In
the case where the `JpxDecodeAction::kConvertArgbToRgb` action is in
use, remove the channel check when selecting the output image format.
Instead, use CHECK_GE() to make sure the code that chose to use
`kConvertArgbToRgb` only did that when there are 4 or more channels.

Bug: 345518608
Change-Id: I3574d82a0d74c6e50da929c6cabb2b0da7ebb208